re-structure
code
