+++ src/router/index.ts
@@ -1,0 +1,11 @@
+import express from 'express';
+
+import test from './test';
+
+const router = express.Router();
+
+export default (): express.Router => {
+    test(router);
+
+    return router;
+}
